fix(architecture): move path helpers to utils to avoid upward dependencies
Move get_vibe3_cache_path, get_vibe3_db_path, and get_vibe3_log_dir
from services.shared.paths (layer 3) to utils.paths (layer 6) to
eliminate upward dependency violations.

Architecture layers:
- Layer 6: utils (infrastructure)
- Layer 5: environment
- Layer 3: services

Problem:
- clients (L6) and environment (L5) were importing from services (L3)
- This violates the layer dependency rule: Layer N may import from layers >= N

Solution:
- Create new utils/paths.py for path helpers
- Update all imports to use vibe3.utils instead of vibe3.services
- Remove helpers from services module

Test results:
- Modularity test passes
- All affected tests pass (47 tests)
- No upward dependency violations

Ref #2514
