+++ src/vibe3/clients/recent_pr_cache.py
@@ -17,7 +17,7 @@ from typing import Any
 
 from loguru import logger
 
-from vibe3.services import get_vibe3_cache_path
+from vibe3.utils import get_vibe3_cache_path
 
 
 class RecentPRCache:
