+++ src/vibe3/services/shared/__init__.py
@@ -23,9 +23,6 @@ if TYPE_CHECKING:
         check_ref_exists,
         find_worktree_path_for_branch,
         get_git_common_dir,
-        get_vibe3_cache_path,
-        get_vibe3_db_path,
-        get_vibe3_log_dir,
         get_worktree_root,
         normalize_ref_path,
         ref_to_handoff_cmd,
